Added time splitting so any time range can now be specified
