+++ sentera/api.py
@@ -1,12 +1,6 @@
 import requests
 from sentera import weather
 from pandas.io.json import json_normalize
-import datetime
-import json
-import pandas as pd
-import aiohttp
-import asyncio
-import tqdm
 
 
 def _run_sentera_query(query, token):
@@ -16,55 +10,6 @@ def _run_sentera_query(query, token):
         raise Exception("Request Failed {}. {}".format(request.status_code, query))
 
     return request.json()
-
-
-async def _fetch(url, session, weather_variable, time_interval, weather_type):
-    num_retries = 0
-    while num_retries < 10:
-        try:
-            async with session.get(url,
-                                   params=weather.create_params(weather_type, time_interval),
-                                   raise_for_status=True) as response:
-                return await response.read(), weather_variable
-        except aiohttp.ClientError as e:
-            # print(e)
-            await asyncio.sleep(1)
-            num_retries += 1
-
-    raise aiohttp.ClientError(f"Couldn't access data from {url} from {time_interval[0]} to {time_interval[1]}")
-
-
-async def _run_weather_queries(url_list, weather_variable_list, time_interval, weather_interval, weather_type):
-    tasks = []
-
-    async with aiohttp.ClientSession(headers=weather.WEATHER_HEADER) as session:
-        for url, weather_variable in zip(url_list, weather_variable_list):
-            task = asyncio.ensure_future(_fetch(url, session, weather_variable, time_interval, weather_type))
-            tasks.append(task)
-
-        data_df = pd.DataFrame(columns=[weather.TIME_COLUMNS[weather_interval], "lat", "long"])
-        for f in tqdm.tqdm(asyncio.as_completed(tasks), total=len(tasks)):
-            response, weather_variable = await f
-            try:
-                response_json = json.loads(response)
-                data = json_normalize(response_json['series'])
-                data = data.rename(columns={"value": str(weather_variable)}).drop(columns=["products"])
-                data["lat"] = response_json['latitude']
-                data["long"] = response_json['longitude']
-
-                data_df = data_df.merge(data,
-                                        on=[weather.TIME_COLUMNS[weather_interval], "lat", "long"],
-                                        how="outer",
-                                        suffixes=["", "_"])
-                if str(weather_variable) + "_" in data_df:
-                    indx = data_df[str(weather_variable) + "_"].notnull()
-                    data_df.loc[indx, str(weather_variable)] = data_df.loc[indx, str(weather_variable) + "_"]
-                    data_df.drop(str(weather_variable) + "_", inplace=True, axis=1)
-            except Exception as e:
-                print(response)
-                raise e
-
-    return data_df
 
 
 def get_all_fields(token):
@@ -97,38 +42,27 @@ async def get_weather(weather_type, weather_variables, weather_interval, locatio
     weather_type = weather.WeatherType(weather_type)
     weather_interval = weather.WeatherInterval(weather_interval)
 
-    if weather_type == weather.WeatherType.Recent:
-        if not time_interval:
-            raise ValueError("Time interval needed for recent weather types")
-
-        try:
-            start = datetime.datetime.strptime(time_interval[0], '%Y/%m/%d')
-        except ValueError:
-            raise ValueError("Incorrect time interval format, should be YYYY/MM/DD")
-
-        try:
-            end = datetime.datetime.strptime(time_interval[1], '%Y/%m/%d')
-        except ValueError:
-            raise ValueError("Incorrect time interval format, should be YYYY/MM/DD")
-
     url_list = []
     weather_variables_list = []
+    time_interval_list = []
 
-    # time_intervals = split_time_interval(start, end, weather_type)
+    time_intervals = weather.split_time_interval(time_interval, weather_type, weather_interval)
 
-    for field_location in location_list:
-        for weather_variable in weather_variables:
-            weather_variable = weather.WeatherVariable(weather_variable)
-            weather_url = weather.build_weather_url(weather_type,
-                                                    weather_variable,
-                                                    weather_interval,
-                                                    field_location[0],
-                                                    field_location[1])
-            url_list.append(weather_url)
-            weather_variables_list.append(weather_variable)
+    for time_interval in time_intervals:
+        for field_location in location_list:
+            for weather_variable in weather_variables:
+                weather_variable = weather.WeatherVariable(weather_variable)
+                weather_url = weather.build_weather_url(weather_type,
+                                                        weather_variable,
+                                                        weather_interval,
+                                                        field_location[0],
+                                                        field_location[1])
+                url_list.append(weather_url)
+                weather_variables_list.append(weather_variable)
+                time_interval_list.append(time_interval)
 
-    return await _run_weather_queries(url_list,
-                                      weather_variables_list,
-                                      time_interval,
-                                      weather_interval,
-                                      weather_type)
+    return await weather.run_queries(url_list,
+                                     weather_variables_list,
+                                     time_interval_list,
+                                     weather_interval,
+                                     weather_type)
